Test-Input Validations for OCC step [x] Removed the Other Considerations spec and separated each substep into individual tests [x]Added Validations for each substep
AT-7389
