+++ cypress/selectors/occ.sel.js
@@ -2,7 +2,8 @@ module.exports = {
   //PII
   blueInfoMessageText:"._atat-alert",
   piiRadioOptionText: "#PIIOptions_radio_group_control legend",
-  piiRadioOtionActive:"#PIIOptions_radio_group_control .v-item--active",
+  piiRadioOtionActive: "#PIIOptions_radio_group_control .v-item--active",
+  piiRadioError:"#PIIOptions_radio_group_control .field-error",
   yesPIIRadioOption: "#Radio_YesPII",
   noPIIRadioOption: "#Radio_NoPII",
   piiLink: "#Button_PIIFAQ",
@@ -13,8 +14,10 @@ module.exports = {
   //PII-Tell Us More about your system of records
   systemLabel: "#SystemName_text_field_label",
   systemNameTextBox: "#SystemName_text_field",
+  systemNameError:"#SystemName_text_field_control .field-error",
   operationPerformedLabel: "#OperationToBePerformed_text_field_label",
-  operationPerformedTextBox:"#OperationToBePerformed_text_area",  
+  operationPerformedTextBox: "#OperationToBePerformed_text_area",
+  operationPerformedError:"#OperationToBePerformed_text_field_control .field-error",
   //BAA
   baaLabelText: "._atat-alert h2",
   phiTextMessage: "._atat-alert._callout p",
@@ -25,11 +28,13 @@ module.exports = {
   noBAARadioOption: "#Radio_NoBAA",
   aboutBALink: "#Button_AboutBusinessAssociates",
   contentAboutBA: "#Content_AboutBusinessAssociates",
+  baaRadioError:"#BAARadioOptions_radio_group_control .field-error",
   moreInfoBAALink: "a._text-link._external-link",
   infoAlert: "._atat-alert._info-alert p",
   //FOIA
   foiaYesOption: "#Radio_Yes",
   foiaNoOption: "#Radio_No",
+  foiaRadioError:"#FOIAOptions_radio_group_control .field-error",
   foiaRadioOptionActive:"#FOIAOptions_radio_group_control .v-item--active",
   foiaLearnLink: "#FOIALearnMore",
   foiaLearnDrawer: ".v-navigation-drawer__content",
@@ -42,9 +47,11 @@ module.exports = {
   //Tell us about your FOIA cordinator
   fullNameLabel: "#FullName_text_field_label",
   fullNameTxtBox: "#FullName_text_field",
+  fullNameError:"#FullName_text_field_control .field-error",
   fullNameHelpTxt: "#FullName_text_field_control .help-text",
   emailLabel: "#Email_text_field_label",
   emailTxtbox: "#Email_text_field",
+  emailError:"#Email_text_field_control .field-error",
   emailHelpTxt: "#Email_text_field_control .help-text",
   //508 standards
   introText: "#IntroP",
